Split the records ontology out of the IPC monolith
Persons, places, events, evidence, relationships, social contacts, kinship
suggestions, the gazetteer and GEDCOM — the ontology the genealogy and
primary-source vaults share.

  electron/ipc.ts      881-1074 -> electron/ipc/records.ts     (64 handlers)
  electron/preload.ts   171-243 -> electron/preload/records.ts (64 bindings)
  shared/types.ts     7347-7434 -> shared/api/records.ts       (64 signatures)

works:scanRecords went with them rather than staying with the corpus: it scans a work
for records and is the only works: channel the records surface owns. Extracting it by
full channel name is what made the block pure in all three files at once — no
relocation needed anywhere.

Where the split stands: 746 of 1,206 channels live in a domain module, 460 remain in
electron/ipc.ts. That file is 2,174 lines (from 6,013 — down 64%),
electron/preload.ts 845 (from 1,862 — down 55%), shared/types.ts 8,133.

Verified: lint and typecheck clean, contract test green, 1,424 of 1,425 tests. The one
failure is test-primary-sources-performance.mjs, which passes standalone and is
timing-sensitive; it covers primarySources: channels, untouched by this change.
